getEventsByVolunteer, Exception handling, minor
Added getVolunteerEventsyByVolunteer to evaluate the return of
(Volunteer).getEventVolunteer(), so a list of events connected to a
volunteer can be displayed/used. Changed FeedPage to update the current
volunteer using the volunteer object returned by
VolunteerService.getVolunteerEventsyByVolunteer(), not sure if the
update is necessary, will depend on how this gets used later. Related to
Issue #10.
Changed the exception thrown in DBConnection to HibernateException so
that exception handling is more specific, removed throws exception from
VolunteerService since the DB queries are surrounded in try/catch blocks
for HibernateException.
Changed some id tags and added some CSS to try to wrangle the layout.
Changed queried volunteer name in VolunteerPage to stop the test
LandingPage from crashing.