navigation, primefaces, template, style, login
Added a navigation menubar to the banner, made more elements primefaces
elements instead of plain-ol'-jsf, fixed some styling issues so that
elements aren't all over the place anymore, removed the banner
replacement in Profile and added the template to Search. Changed login
so that it redirects, instead of forwards, to the current page,
eliminating the "resend information" on reload, and displaying the
correct URI in the browser.