Rewrite integration
Added rewrite integration, related to #18, #21.
Added some join rules to FacesConfigurationProvider,
http://ocpsoft.org/rewrite/docs/ has more information on how to create
rules.
Updated local Tomcat version to 7.0.50, this fixed some issues with
integrating Rewrite, but exposed poor practices in coding in the android
servlets: Tomcat 7.0.50 will not load a WAR with multiple servlets
mapped to the same URL, so I changed the URL mappings in the android
interface servlets to be "/Android/get*" where * was the name of the
resource being gotten.
Changed web.xml to map only *.xhtml pages to the Faces Servlet, removed
the welcom-file-list, this was required to enable Rewrite to work
correctly with Faces.